refactor: remove bad common
