+++ src/main/java/app/bottlenote/review/dto/response/ReviewDetailResponse.java
@@ -1,19 +1,16 @@
 package app.bottlenote.review.dto.response;
 
 import app.bottlenote.alcohols.dto.response.AlcoholInfo;
-import app.bottlenote.review.dto.common.CommonReviewInfo;
 import app.bottlenote.review.dto.request.ReviewImageInfo;
+import app.bottlenote.review.dto.vo.CommonReviewInfo;
 
 import java.util.List;
 
 public record ReviewDetailResponse(
 	AlcoholInfo alcoholInfo,
-
 	CommonReviewInfo reviewResponse,
-
 	List<ReviewImageInfo> reviewImageList
 ) {
-
 	public static ReviewDetailResponse create(AlcoholInfo alcoholInfo, CommonReviewInfo reviewResponse, List<ReviewImageInfo> reviewImageList) {
 		if (reviewResponse == null) {
 			return new ReviewDetailResponse(null, null, null);
@@ -21,4 +18,3 @@ public record ReviewDetailResponse(
 		return new ReviewDetailResponse(alcoholInfo, reviewResponse, reviewImageList);
 	}
 }
-
