+++ src/main/java/app/bottlenote/review/dto/response/ReviewListResponse.java
@@ -2,16 +2,16 @@ package app.bottlenote.review.dto.response;
 
 import app.bottlenote.review.domain.constant.ReviewDisplayStatus;
 import app.bottlenote.review.domain.constant.SizeType;
+import lombok.Builder;
+
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
-import lombok.Builder;
 
 public record ReviewListResponse(
 	Long totalCount,
 	List<ReviewInfo> reviewList
 ) {
-
 	public static ReviewListResponse of(Long totalCount, List<ReviewInfo> reviewList) {
 		return new ReviewListResponse(totalCount, reviewList);
 	}
