+++ src/main/java/app/bottlenote/review/dto/vo/UserInfo.java
@@ -1,4 +1,4 @@
-package app.bottlenote.review.dto.common;
+package app.bottlenote.review.dto.vo;
 
 public record UserInfo(
 	Long userId,
