+++ src/main/java/app/bottlenote/review/repository/custom/CustomReviewQueryRepository.java
@@ -2,7 +2,7 @@ package app.bottlenote.review.repository.custom;
 
 
 import app.bottlenote.alcohols.dto.response.detail.ReviewsDetailInfo;
-import app.bottlenote.review.dto.common.CommonReviewInfo;
+import app.bottlenote.review.dto.vo.CommonReviewInfo;
 
 import java.util.List;
