+++ src/main/java/app/bottlenote/review/repository/custom/CustomReviewRepository.java
@@ -1,9 +1,9 @@
 package app.bottlenote.review.repository.custom;
 
 import app.bottlenote.global.service.cursor.PageResponse;
-import app.bottlenote.review.dto.common.CommonReviewInfo;
 import app.bottlenote.review.dto.request.ReviewPageableRequest;
 import app.bottlenote.review.dto.response.ReviewListResponse;
+import app.bottlenote.review.dto.vo.CommonReviewInfo;
 
 public interface CustomReviewRepository {
